+++ core-model/src/test/kotlin/com/exposures/model/FrameNumberingTest.kt
@@ -12,6 +12,7 @@ class FrameNumberingTest {
         filmRollId = "roll-1",
         frameNumber = frameNumber,
         lensId = "lens-1",
+        focalLengthMm = null,
         shutterSpeed = ShutterSpeed.fraction(125),
         aperture = 8.0,
         isoUsed = 400,